bot/requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
Deprecated==1.2.13
et-xmlfile==1.1.0
fire==0.4.0
frozenlist==1.3.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==3.3.0
multidict==6.0.2
openpyxl==3.0.10
psycopg2==2.9.3
pyattck==6.1.2
pyattck-data==2.4.2
pycparser==2.21
pydantic==1.9.1
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
termcolor==1.1.0
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
